+++ CHANGELOG.md
@@ -11,6 +11,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
     - config_key from ``src/Codeception/Template/ModuleAcceptance/acceptance.suite.yml``
     - protected requiredFields in ``src/Module/Database.php``
 
+## [1.2.0] - Unreleased
+
+### Added
+- Use declare(strict_types=1); in template files
+- Screen shot url for failing tests
+
+### Fix
+- Fix bootstrap's template configuration with ^3.1 codeception version
+
 ## [1.1.0] -  2019-11-07
 
 ### Added